# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ashva Experts API (Consolidated)
  version: 0.2.0
  description: |
    Consolidated API specification for manual testing during development.

    Notes:
    - Base URL for local testing via Nginx: `http://localhost:8080`
    - All `/api/v1/*` routes are served through the gateway.
    - Auth endpoints under `/api/v1/auth/*` are public.
    - Non-auth endpoints require `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:8080
    description: Local via Nginx (recommended)
  - url: http://localhost:8000
    description: Local gateway direct
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
    RegisterRequest:
      type: object
      properties:
        email: { type: string, format: email, example: user@example.com }
        phone: { type: string, example: "+919876543210" }
        password: { type: string, example: "SecurePassword123!" }
        role:
          type: string
          enum: [admin, subscriber, technician, cms_user]
          example: subscriber
        first_name: { type: string, example: John }
        last_name: { type: string, example: Doe }
      required: [email, password, role]
    RegisterResponse:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        message: { type: string }
      required: [user_id, email, message]
    OtpRequestRequest:
      type: object
      properties:
        identifier:
          type: string
          description: Email or phone
          example: user@example.com
        purpose:
          type: string
          enum: [login, registration, password_reset]
          example: login
      required: [identifier, purpose]
    OtpRequestResponse:
      type: object
      properties:
        message: { type: string, example: OTP sent successfully }
        expires_in_seconds: { type: integer, example: 300 }
      required: [message, expires_in_seconds]
    OtpVerifyRequest:
      type: object
      properties:
        identifier: { type: string, example: user@example.com }
        otp_code:
          type: string
          description: |
            Local dev uses a static OTP (default `123456`).
          example: "123456"
        purpose:
          type: string
          enum: [login, registration, password_reset]
          example: login
      required: [identifier, otp_code, purpose]
    TokenUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role:
          type: string
          enum: [admin, subscriber, technician, cms_user]
        subscriber_id: { type: string, format: uuid, nullable: true }
        can_assign_leads: { type: boolean, example: false }
        can_manage_unassigned_leads: { type: boolean, example: true }
      required: [id, email, role]
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 3600 }
        user: { $ref: "#/components/schemas/TokenUser" }
      required: [access_token, refresh_token, token_type, expires_in, user]
    PasswordLoginRequest:
      type: object
      properties:
        email: { type: string, format: email, example: user@example.com }
        password: { type: string, example: "SecurePassword123!" }
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token: { type: string }
      required: [refresh_token]
    RefreshResponse:
      type: object
      properties:
        access_token: { type: string }
        expires_in: { type: integer, example: 3600 }
        token_type: { type: string, example: Bearer }
      required: [access_token, expires_in, token_type]
    LogoutRequest:
      type: object
      properties:
        refresh_token: { type: string }
      required: [refresh_token]
    ValidateRequest:
      type: object
      properties:
        token: { type: string }
      required: [token]
    ValidateResponse:
      type: object
      properties:
        valid: { type: boolean, example: true }
        user_id: { type: string, format: uuid }
        email: { type: string, format: email, example: user@example.com }
        role:
          type: string
          enum: [admin, subscriber, technician, cms_user]
        subscriber_id: { type: string, format: uuid, nullable: true }
        can_assign_leads: { type: boolean, example: false }
        can_manage_unassigned_leads: { type: boolean, example: true }
        expires_at: { type: string, format: date-time }
      required: [valid, user_id, email, role, expires_at]
    MeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        role:
          type: string
          enum: [admin, subscriber, technician, cms_user]
        subscriber_id: { type: string, format: uuid, nullable: true }
        can_assign_leads: { type: boolean, example: false }
        can_manage_unassigned_leads: { type: boolean, example: true }
        is_active: { type: boolean }
        is_verified: { type: boolean }
      required: [id, email, role, is_active, is_verified]
...
    StaffCreateRequest:
      type: object
      properties:
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        password: { type: string }
        role: { type: string, enum: [cms_user, technician] }
      required: [email, password, role]
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a user (OTP required)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
            example:
              email: user@example.com
              phone: "+919876543210"
              password: "SecurePassword123!"
              role: subscriber
              first_name: John
              last_name: Doe
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "409":
          description: Already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/otp/request:
    post:
      tags: [auth]
      summary: Request OTP
      description: |
        Local dev uses a static OTP (default `123456`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpRequestRequest" }
            example:
              identifier: user@example.com
              purpose: login
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OtpRequestResponse" }
  /api/v1/auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpVerifyRequest" }
            example:
              identifier: user@example.com
              otp_code: "123456"
              purpose: login
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Invalid OTP
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Password login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordLoginRequest" }
            example:
              email: user@example.com
              password: "SecurePassword123!"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshResponse" }
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout (revoke refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: Logged out successfully
  /api/v1/auth/validate:
    post:
      tags: [auth]
      summary: Validate access token (used by gateway)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ValidateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidateResponse" }
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
...
  /api/v1/auth/subscribers/{subscriber_id}/staff:
    post:
      tags: [auth]
      summary: Create staff user under subscriber tenant (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subscriber_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffCreateRequest" }
      responses:
        "201":
          description: Created
...